+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Enhancements
 - NOTICKET - update changelog, post-release
 ### Fixed bugs
+- GP2-2886 - DAC_Focus_Order_03 - EP Objectives list focus management
 - GP2-2885 - DAC_Focus_Order_02 - Dynamic aria label and focus on WTM page
 - GP2-2884 - DAC Focus Order 01 - tabindex on mobile header
 - GP2-2893 - DAC Link Purpose 01